Tiger: fix MSVC warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk@31604 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/tiger/tigercompletechain.cpp
@@ -516,8 +516,8 @@ int TigerCompleteChain::AddShapePoints( int nTLID, int nRecordId,
             return FALSE;
         }
 
-        nBytesRead = VSIFReadL( achShapeRec, 1, psRT2Info->nRecordLength, 
-                                fpShape );
+        nBytesRead = static_cast<int>(VSIFReadL( achShapeRec, 1, psRT2Info->nRecordLength, 
+                                fpShape ));
 
         /* 
         ** Handle case where the last record in the file is full.  We will
